Animated bubbles and fixed timer
All of the birds' "speech bubbles" are now animated to fade in. This provides a visual cue in addition to their sounds. The hungry bird's bubble still shows the worm. A famished bird's bubble alternates between a worm and a "hangry" icon.

Next step will be to do a similar alternating animation for the randy/mating bubble. It will alternate between a heart and the nesting materials.

I also extended the timer on the Title Screen to 4 seconds. The "tutorial arrow" will not start until then and will not repeat until four seconds after the animation has finished playing.