Working on high scores. Bugs remain.
In progress.

I changed the Global.high_scores variable to a nested array (list of lists), and it sorts better now.

New high scores are added to the table, but there is a problem. The manage_high_scores() function always runs exactly twice, causing a duplicate entry of a new high score in some cases. I suspect this is due to something that happens when the lose_screen _ready() function is called.