Fixed high scores but they don't save to file
On game over, the "You Flocked Up" text animation appears. If the player's score is higher than the lowest, high score, the player is prompted to enter their name. (They have 10 seconds to do so or the name "TooSlow..." is entered instead.) Then, the game transitions to the game over screen, showing the table of high scores.

Currently, the new high scores do not write to a file. So, they reset to defaults each time the game is run.