High Scores table completely functional
The table of high scores now writes to a text file on the user's system. During subsequent launches, the user's file is used to populate the initial table of high scores rather than the default table in the res:// files. For reference, the location of the file on my Windows machine is:
C:\Users\User\AppData\Roaming\Godot\app_userdata\Point-and-Flock